# docs/api/openapi-specification.yaml
openapi: 3.0.0
info:
  title: Portfolio Site REST API
  version: 2.0.0
  description: |
    Comprehensive REST API for Hugo-based portfolio site with content management,
    analytics, webhooks, and GraphQL integration.
    
    ## Features
    - Full CRUD operations for all content types
    - Advanced search and filtering
    - Real-time analytics and reporting
    - Webhook system for integrations
    - GraphQL endpoint for flexible queries
    - Multi-language content support
    - Image optimization and management
    - SEO and metadata management
    
  contact:
    name: API Support
    email: support@portfolio.com
    url: https://portfolio.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.portfolio.com/v2
    description: Production server
  - url: https://staging-api.portfolio.com/v2
    description: Staging server
  - url: http://localhost:3000/api/v2
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /search:
    get:
      tags: [Search]
      summary: Global content search
      description: Advanced search across all content types with full-text search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Search query
        - name: type
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [blog, portfolio, tools, teaching, pages]
          description: Content types to search
        - name: language
          in: query
          schema:
            type: string
            enum: [en, es]
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, date, title, views]
            default: relevance
        - name: dateRange
          in: query
          schema:
            type: string
            enum: [week, month, year, all]
            default: all
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
  /search/suggestions:
    get:
      tags: [Search]
      summary: Get search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        '200':
          description: Search suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    text:
                      type: string
                    type:
                      type: string
                    count:
                      type: integer

  # Analytics and Reporting
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    SearchResults:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        facets:
          type: object
          properties:
            types:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
            categories:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
            tags:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  count:
                    type: integer
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    SearchResult:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [blog, portfolio, tools, teaching, pages]
        title:
          type: string
        excerpt:
          type: string
        url:
          type: string
        language:
          type: string
        publishedAt:
          type: string
          format: date-time
        relevanceScore:
          type: number
        highlights:
          type: array
          items:
            type: string
...